backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
Pillow==9.0.0
djoser==2.2.2
django-filter==23.5